+++ Product.h
@@ -10,11 +10,11 @@ class Product
 private:
 	size_t productCode;
 	string productName;
-	size_t unitCentPrice; // price of one item of the product WITH cents/copeiks (*100)
+	long long unitCentPrice; // price of one item of the product WITH cents/copeiks (*100)
 	string manufacturer;  // name of the manufacturer of the product
-	size_t weightNet;     // net weight of one item of the product in grams
-	size_t weightGross;   // gross weight of one item of the product in grams
-	size_t width, height, length;   // dimensions in millimiters
+	long long weightNet;     // net weight of one item of the product in grams
+	long long weightGross;   // gross weight of one item of the product in grams
+	long long width, height, length;   // dimensions in millimiters
 
 /* The following functionality is removed	
 	//bool soldByWeight;    // shows if the product is sold by weight (without fixed package)
@@ -27,23 +27,23 @@ public:
 	Product(
 		size_t productCode,
 		string productName,
-		size_t unitCentPrice,
+		long long unitCentPrice,
 		string manufacturer = "",
-		size_t weightNet = 0,
-		size_t weightGross = 0,
-		size_t width = 0,
-		size_t height = 0, 
-		size_t length = 0
+		long long weightNet = 0,
+		long long weightGross = 0,
+		long long width = 0,
+		long long height = 0, 
+		long long length = 0
 		);
 
 	size_t GetCode();
 	string GetName();
-	size_t GetNetWeight();
-	size_t GetGrossWeight();
-	size_t GetWidth();
-	size_t GetHeight();
-	size_t GetLength();
-	size_t GetUnitCentPrice();
+	long long GetNetWeight();
+	long long GetGrossWeight();
+	long long GetWidth();
+	long long GetHeight();
+	long long GetLength();
+	long long GetUnitCentPrice();
 
 	// Returns productCode and productName as a string looking like '(code) "name"'
 	string GetCodeAndName(); 
@@ -54,8 +54,8 @@ public:
 
 /* The following functionality is removed	
 	//void SetManufacturer(string manufacturer);
-	//void SetWeightGross(size_t weightGross);
-	//void SetDimensions(size_t width, size_t height, size_t length);
+	//void SetWeightGross(long long weightGross);
+	//void SetDimensions(long long width, long long height, long long length);
 	//void SetProductionDate(int prod_mday, int prod_month, int prod_year); // takes: day of month from 1 to 31, month of year from 1 to 12, year
 	//void SetExpirationDate(int exp_mday, int exp_mon, int exp_year); // takes: day of month from 1 to 31, month of year from 1 to 12, year
 	//string GetProductionDate();  // returns a string "DD/MM/YYYY"
